+++ src/test/java/org/jjmirowski/model/MatchTest.java
@@ -1,23 +1,41 @@
 package org.jjmirowski.model;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 class MatchTest {
 
+    private String homeTeam;
+    private String awayTeam;
+
+    @BeforeEach
+    void setUp() {
+        homeTeam = "Poland";
+        awayTeam = "Italy";
+    }
+
     @Test
     void shouldCreateMatchWithInitialScoreZero() {
-        Match match = new Match("Poland", "Italy");
-        assertEquals("Poland", match.getHomeTeam());
-        assertEquals("Italy", match.getAwayTeam());
+        // when
+        Match match = new Match(homeTeam, awayTeam);
+
+        //then
+        assertEquals(homeTeam, match.getHomeTeam());
+        assertEquals(awayTeam, match.getAwayTeam());
         assertEquals(0, match.getHomeScore());
         assertEquals(0, match.getAwayScore());
     }
 
     @Test
     void shouldUpdateScore() {
-        Match match = new Match("Poland", "Italy");
+        // given
+        Match match = new Match(homeTeam, awayTeam);
+
+        // when
+        match.updateScore(4, 0);
+
         match.updateScore(4, 0);
         assertEquals(4, match.getHomeScore());
         assertEquals(0, match.getAwayScore());